requirements.txt
python-dotenv
pandas
requests
chromadb 
langchain-core 
langchain-chroma 
langchain-ollama
langchain-community
unstructured[md]
nltk
faiss-cpu
faiss-cpu
logging
